authorizarion done

+++ app/Http/Controllers/AdminsController.php
@@ -5,13 +5,22 @@ namespace App\Http\Controllers;
 use Exception;
 use App\Models\Admin;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Spatie\Permission\Models\Permission;
 
 class AdminsController extends Controller
 {
+    public $user;
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -19,6 +28,9 @@ class AdminsController extends Controller
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('admin.view')) {
+            abort(403, 'You are Unauthorized !');
+        }
         $admins = Admin::all();
         return view('backend.pages.admins.index', compact('admins'));
     }
@@ -30,6 +42,9 @@ class AdminsController extends Controller
      */
     public function create()
     {
+        if (is_null($this->user) || !$this->user->can('admin.create')) {
+            abort(403, 'You are Unauthorized !');
+        }
         $roles  = Role::all();
         return view('backend.pages.admins.create', compact('roles'));
     }
@@ -42,6 +57,9 @@ class AdminsController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('admin.create')) {
+            abort(403, 'You are Unauthorized !');
+        }
         // Validation Data
         $request->validate([
             'name' => 'required|max:50',
@@ -85,6 +103,9 @@ class AdminsController extends Controller
      */
     public function edit($id)
     {
+        if (is_null($this->user) || !$this->user->can('admin.edit')) {
+            abort(403, 'You are Unauthorized !');
+        }
         $admin = Admin::find($id);
         $roles  = Role::all();
         return view('backend.pages.admins.edit', compact('admin', 'roles'));
@@ -99,6 +120,9 @@ class AdminsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (is_null($this->user) || !$this->user->can('admin.edit')) {
+            abort(403, 'You are Unauthorized !');
+        }
         // find  Admin
         $admin = Admin::find($id);
 
@@ -136,6 +160,9 @@ class AdminsController extends Controller
      */
     public function destroy($id)
     {
+        if (is_null($this->user) || !$this->user->can('admin.delete')) {
+            abort(403, 'You are Unauthorized !');
+        }
         $admin = Admin::find($id);
         if (!is_null($admin)) {
             $admin->delete();
